Add download-impl command to lb-steve cli. Example usage:
  $ lb-steve download-impl -o asd.tar.gz --impl noop

When using a local filename, behavior is the same as the -o option of create-job,
which means it gets downloaded to the local machine. For this to work the
config file needs to have a default_output_prefix (s3 url) set, that allows
writing by LB Jobs, as well as reading by the lb-steve client.

Implemented a retry for any exception, except the ServiceClientException/SteveClientException